pyproject.toml
[tool.poetry]
name = "book_summarizer"
version = "0.1.0"
description = ""
authors = ["Your Name <your email>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.16"
openai = "^1.58.1"
python-dotenv = "^1.0.1"
langchain = "^0.3.14"
chroma = "^0.2.0"
langchain-community = "^0.3.14"
chromadb = "^0.6.1"
langchain-openai = "^0.2.14"
flask = "^3.1.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pandas = "^2.2.3"
faiss-cpu = "^1.9.0.post1"
black = "^24.10.0"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
